Add jest test coverage
