+++ bundles/@yarnpkg/plugin-licenses.js
@@ -2,7 +2,7 @@
 module.exports = {
 name: "@yarnpkg/plugin-licenses",
 factory: function (require) {
-var plugin;plugin=(()=>{"use strict";var __webpack_modules__={518:(e,_,r)=>{r.r(_),r.d(_,{default:()=>k});var t={};r.r(t),r.d(t,{getPackageManifest:()=>l});const n=require("@yarnpkg/cli");var i=r(966);const a=require("clipanion");var o=r(300);const s=require("@yarnpkg/parsers"),c=require("fs");var p=r(688);const l=(e,_)=>{d(e);const r=i.structUtils.convertPackageToLocator(_),t=u[i.structUtils.stringifyLocator(r)];if(!t)return null;const n=t.locations[0],a=n?p.ppath.join(n,p.Filename.manifest):p.Filename.manifest,o=p.npath.fromPortablePath(a),s=(0,c.readFileSync)(o).toString();return JSON.parse(s)};let u;const d=e=>{if(!u){const _=p.ppath.join(e.configuration.projectCwd,p.Filename.nodeModules,".yarn-state.yml"),r=p.npath.fromPortablePath(_);u=(0,s.parseSyml)((0,c.readFileSync)(r).toString())}};var f=function(e,_,r,t){var n,i=arguments.length,a=i<3?_:null===t?t=Object.getOwnPropertyDescriptor(_,r):t;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)a=Reflect.decorate(e,_,r,t);else for(var o=e.length-1;o>=0;o--)(n=e[o])&&(a=(i<3?n(a):i>3?n(_,r,a):n(_,r))||a);return i>3&&a&&Object.defineProperty(_,r,a),a};class g extends a.Command{constructor(){super(...arguments),this.recursive=!1,this.json=!1}async execute(){const e=await i.Configuration.find(this.context.cwd,this.context.plugins),{project:_,workspace:r}=await i.Project.find(e,this.context.cwd);if(!r)throw new n.WorkspaceRequiredError(_.cwd,this.context.cwd);await _.restoreInstallState();const t=await P(_,this.json,this.recursive);i.treeUtils.emitTree(t,{configuration:e,stdout:this.context.stdout,json:this.json,separators:1})}}g.usage=a.Command.Usage({description:"display the licenses for all packages in the project",details:"\n      This command prints the license information for packages in the project. By default, only direct dependencies are listed.\n\n      If `-R,--recursive` is set, the listing will include transitive dependencies (dependencies of direct dependencies).\n    ",examples:[["List all licenses of direct dependencies","$0 licenses list"],["List all licenses of direct and transitive dependencies","$0 licenses list --recursive"]]}),f([a.Command.Boolean("-R,--recursive")],g.prototype,"recursive",void 0),f([a.Command.Boolean("--json")],g.prototype,"json",void 0),f([a.Command.Path("licenses","list")],g.prototype,"execute",null);const k={commands:[g]},P=async(e,_,r)=>{const n={},a={children:n};let s;s=r?e.storedDescriptors.values():e.workspaces.flatMap(e=>{const _=[e.anchoredDescriptor];return _.push(...e.dependencies.values()),_});const c=i.miscUtils.sortMap(s,e=>i.structUtils.stringifyDescriptor(e)),p=(e=>{switch(e){case"pnp":return o;case"node-modules":return t;default:throw new Error("Unsupported linker")}})(e.configuration.get("nodeLinker"));for(const r of c.values()){const t=e.storedResolutions.get(r.descriptorHash);if(!t)continue;const a=e.storedPackages.get(t);if(!a)continue;const o=i.structUtils.convertPackageToLocator(a),s=p.getPackageManifest(e,a);if(null===s)continue;const{license:c,url:l,vendorName:u,vendorUrl:d}=y(s);n[c]||(n[c]={value:i.formatUtils.tuple(i.formatUtils.Type.NO_HINT,c),children:{}});const f={value:i.formatUtils.tuple(i.formatUtils.Type.DEPENDENT,{locator:o,descriptor:r}),children:{...l?{url:{value:i.formatUtils.tuple(i.formatUtils.Type.NO_HINT,E("URL",l,_))}}:{},...u?{vendorName:{value:i.formatUtils.tuple(i.formatUtils.Type.NO_HINT,E("VendorName",u,_))}}:{},...d?{vendorUrl:{value:i.formatUtils.tuple(i.formatUtils.Type.NO_HINT,E("VendorUrl",d,_))}}:{}}},g=i.structUtils.stringifyLocator(o);n[c].children[g]=f}return a},y=e=>{const{license:_,repository:r,homepage:t,author:n}=e;return{license:("string"!=typeof _?null==_?void 0:_.type:_)||"UNKNOWN",url:(null==r?void 0:r.url)||t,vendorName:null==n?void 0:n.name,vendorUrl:t||(null==n?void 0:n.url)}},E=(e,_,r)=>r?_:`${e}: ${_}`},300:(__unused_webpack_module,__webpack_exports__,__webpack_require__)=>{__webpack_require__.r(__webpack_exports__),__webpack_require__.d(__webpack_exports__,{getPackageManifest:()=>getPackageManifest});var _yarnpkg_plugin_pnp__WEBPACK_IMPORTED_MODULE_0__=__webpack_require__(798),_yarnpkg_plugin_pnp__WEBPACK_IMPORTED_MODULE_0___default=__webpack_require__.n(_yarnpkg_plugin_pnp__WEBPACK_IMPORTED_MODULE_0__),_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1__=__webpack_require__(966),_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1___default=__webpack_require__.n(_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1__),_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__=__webpack_require__(688),_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2___default=__webpack_require__.n(_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__),_yarnpkg_libzip__WEBPACK_IMPORTED_MODULE_3__=__webpack_require__(76),_yarnpkg_libzip__WEBPACK_IMPORTED_MODULE_3___default=__webpack_require__.n(_yarnpkg_libzip__WEBPACK_IMPORTED_MODULE_3__);const getPackageManifest=(e,_)=>{makePnPApi(e);const r=_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1__.structUtils.convertPackageToLocator(_),t={name:_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1__.structUtils.stringifyIdent(r),reference:r.reference},n=pnpApi.getPackageInformation(t);if(!n)return null;const{packageLocation:i}=n,a=_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__.ppath.join(i,_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__.Filename.manifest),o=fs.readFileSync(a).toString();return JSON.parse(o)};let pnpApi;const makePnPApi=project=>{pnpApi||(pnpApi=eval("module.require")((0,_yarnpkg_plugin_pnp__WEBPACK_IMPORTED_MODULE_0__.getPnpPath)(project).main))},fs=new _yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__.VirtualFS({baseFs:new _yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__.ZipOpenFS({libzip:(0,_yarnpkg_libzip__WEBPACK_IMPORTED_MODULE_3__.getLibzipSync)(),readOnlyArchives:!0})})},966:e=>{e.exports=require("@yarnpkg/core")},688:e=>{e.exports=require("@yarnpkg/fslib")},76:e=>{e.exports=require("@yarnpkg/libzip")},798:e=>{e.exports=require("@yarnpkg/plugin-pnp")}},__webpack_module_cache__={};function __webpack_require__(e){if(__webpack_module_cache__[e])return __webpack_module_cache__[e].exports;var _=__webpack_module_cache__[e]={exports:{}};return __webpack_modules__[e](_,_.exports,__webpack_require__),_.exports}return __webpack_require__.n=e=>{var _=e&&e.__esModule?()=>e.default:()=>e;return __webpack_require__.d(_,{a:_}),_},__webpack_require__.d=(e,_)=>{for(var r in _)__webpack_require__.o(_,r)&&!__webpack_require__.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:_[r]})},__webpack_require__.o=(e,_)=>Object.prototype.hasOwnProperty.call(e,_),__webpack_require__.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},__webpack_require__(518)})();
+var plugin;plugin=(()=>{"use strict";var __webpack_modules__={559:(e,_,r)=>{r.r(_),r.d(_,{default:()=>m});var t={};r.r(t),r.d(t,{getPackageManifest:()=>l});const n=require("@yarnpkg/cli");var a=r(966);const i=require("clipanion");var o=r(688),s=r(300);const c=require("@yarnpkg/parsers"),p=require("fs"),l=(e,_)=>{d(e);const r=a.structUtils.convertPackageToLocator(_),t=u[a.structUtils.stringifyLocator(r)];if(!t)return null;const n=t.locations[0],i=n?o.ppath.join(n,o.Filename.manifest):o.Filename.manifest,s=o.ppath.join(e.cwd,i),c=o.npath.fromPortablePath(s),l=(0,p.readFileSync)(c).toString();return JSON.parse(l)};let u;const d=e=>{if(!u){const _=o.ppath.join(e.configuration.projectCwd,o.Filename.nodeModules,".yarn-state.yml"),r=o.npath.fromPortablePath(_);u=(0,c.parseSyml)((0,p.readFileSync)(r).toString())}},f=("@yarnpkg"===o.npath.basename(__dirname)?o.ppath.join(o.npath.toPortablePath(__dirname),"..",".."):o.ppath.join(o.npath.toPortablePath(__dirname),".."),async(e,_,r)=>{const n={},i={children:n};let o;o=r?e.storedDescriptors.values():e.workspaces.flatMap(e=>{const _=[e.anchoredDescriptor];return _.push(...e.dependencies.values()),_});const c=a.miscUtils.sortMap(o,e=>a.structUtils.stringifyDescriptor(e)),p=(e=>{switch(e){case"pnp":return s;case"node-modules":return t;default:throw new Error("Unsupported linker")}})(e.configuration.get("nodeLinker"));for(const r of c.values()){const t=e.storedResolutions.get(r.descriptorHash);if(!t)continue;const i=e.storedPackages.get(t);if(!i)continue;const o=a.structUtils.convertPackageToLocator(i),s=p.getPackageManifest(e,i);if(null===s)continue;const{license:c,url:l,vendorName:u,vendorUrl:d}=g(s);n[c]||(n[c]={value:a.formatUtils.tuple(a.formatUtils.Type.NO_HINT,c),children:{}});const f={value:a.formatUtils.tuple(a.formatUtils.Type.DEPENDENT,{locator:o,descriptor:r}),children:{...l?{url:{value:a.formatUtils.tuple(a.formatUtils.Type.NO_HINT,k("URL",l,_))}}:{},...u?{vendorName:{value:a.formatUtils.tuple(a.formatUtils.Type.NO_HINT,k("VendorName",u,_))}}:{},...d?{vendorUrl:{value:a.formatUtils.tuple(a.formatUtils.Type.NO_HINT,k("VendorUrl",d,_))}}:{}}},P=a.structUtils.stringifyLocator(o);n[c].children[P]=f}return i}),g=e=>{const{license:_,repository:r,homepage:t,author:n}=e;return{license:("string"!=typeof _?null==_?void 0:_.type:_)||"UNKNOWN",url:(null==r?void 0:r.url)||t,vendorName:null==n?void 0:n.name,vendorUrl:t||(null==n?void 0:n.url)}},k=(e,_,r)=>r?_:`${e}: ${_}`;var P=function(e,_,r,t){var n,a=arguments.length,i=a<3?_:null===t?t=Object.getOwnPropertyDescriptor(_,r):t;if("object"==typeof Reflect&&"function"==typeof Reflect.decorate)i=Reflect.decorate(e,_,r,t);else for(var o=e.length-1;o>=0;o--)(n=e[o])&&(i=(a<3?n(i):a>3?n(_,r,i):n(_,r))||i);return a>3&&i&&Object.defineProperty(_,r,i),i};class y extends i.Command{constructor(){super(...arguments),this.recursive=!1,this.json=!1}async execute(){const e=await a.Configuration.find(this.context.cwd,this.context.plugins),{project:_,workspace:r}=await a.Project.find(e,this.context.cwd);if(!r)throw new n.WorkspaceRequiredError(_.cwd,this.context.cwd);await _.restoreInstallState();const t=await f(_,this.json,this.recursive);a.treeUtils.emitTree(t,{configuration:e,stdout:this.context.stdout,json:this.json,separators:1})}}y.usage=i.Command.Usage({description:"display the licenses for all packages in the project",details:"\n      This command prints the license information for packages in the project. By default, only direct dependencies are listed.\n\n      If `-R,--recursive` is set, the listing will include transitive dependencies (dependencies of direct dependencies).\n    ",examples:[["List all licenses of direct dependencies","$0 licenses list"],["List all licenses of direct and transitive dependencies","$0 licenses list --recursive"]]}),P([i.Command.Boolean("-R,--recursive")],y.prototype,"recursive",void 0),P([i.Command.Boolean("--json")],y.prototype,"json",void 0),P([i.Command.Path("licenses","list")],y.prototype,"execute",null);const m={commands:[y]}},300:(__unused_webpack_module,__webpack_exports__,__webpack_require__)=>{__webpack_require__.r(__webpack_exports__),__webpack_require__.d(__webpack_exports__,{getPackageManifest:()=>getPackageManifest});var _yarnpkg_plugin_pnp__WEBPACK_IMPORTED_MODULE_0__=__webpack_require__(798),_yarnpkg_plugin_pnp__WEBPACK_IMPORTED_MODULE_0___default=__webpack_require__.n(_yarnpkg_plugin_pnp__WEBPACK_IMPORTED_MODULE_0__),_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1__=__webpack_require__(966),_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1___default=__webpack_require__.n(_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1__),_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__=__webpack_require__(688),_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2___default=__webpack_require__.n(_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__),_yarnpkg_libzip__WEBPACK_IMPORTED_MODULE_3__=__webpack_require__(76),_yarnpkg_libzip__WEBPACK_IMPORTED_MODULE_3___default=__webpack_require__.n(_yarnpkg_libzip__WEBPACK_IMPORTED_MODULE_3__);const getPackageManifest=(e,_)=>{makePnPApi(e);const r=_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1__.structUtils.convertPackageToLocator(_),t={name:_yarnpkg_core__WEBPACK_IMPORTED_MODULE_1__.structUtils.stringifyIdent(r),reference:r.reference},n=pnpApi.getPackageInformation(t);if(!n)return null;const{packageLocation:a}=n,i=_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__.ppath.join(a,_yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__.Filename.manifest),o=fs.readFileSync(i).toString();return JSON.parse(o)};let pnpApi;const makePnPApi=project=>{pnpApi||(pnpApi=eval("module.require")((0,_yarnpkg_plugin_pnp__WEBPACK_IMPORTED_MODULE_0__.getPnpPath)(project).main))},fs=new _yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__.VirtualFS({baseFs:new _yarnpkg_fslib__WEBPACK_IMPORTED_MODULE_2__.ZipOpenFS({libzip:(0,_yarnpkg_libzip__WEBPACK_IMPORTED_MODULE_3__.getLibzipSync)(),readOnlyArchives:!0})})},966:e=>{e.exports=require("@yarnpkg/core")},688:e=>{e.exports=require("@yarnpkg/fslib")},76:e=>{e.exports=require("@yarnpkg/libzip")},798:e=>{e.exports=require("@yarnpkg/plugin-pnp")}},__webpack_module_cache__={};function __webpack_require__(e){if(__webpack_module_cache__[e])return __webpack_module_cache__[e].exports;var _=__webpack_module_cache__[e]={exports:{}};return __webpack_modules__[e](_,_.exports,__webpack_require__),_.exports}return __webpack_require__.n=e=>{var _=e&&e.__esModule?()=>e.default:()=>e;return __webpack_require__.d(_,{a:_}),_},__webpack_require__.d=(e,_)=>{for(var r in _)__webpack_require__.o(_,r)&&!__webpack_require__.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:_[r]})},__webpack_require__.o=(e,_)=>Object.prototype.hasOwnProperty.call(e,_),__webpack_require__.r=e=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},__webpack_require__(559)})();
 return plugin;
 }
 };
