--- src/__tests__/fixtures/test-package-node-modules/.yarnrc.yml
@@ -1,4 +1,0 @@
-nodeLinker: node-modules
-
-plugins:
-  - ../../../../bundles/@yarnpkg/plugin-licenses.js
