--- src/__tests__/fixtures/test-package-pnp/.yarnrc.yml
@@ -1,3 +1,0 @@
-
-plugins:
-  - ../../../../bundles/@yarnpkg/plugin-licenses.js
