+++ src/__tests__/integration/fixtures/test-package-node-modules/.yarnrc.yml
@@ -1,0 +1,6 @@
+enableColors: false
+
+nodeLinker: node-modules
+
+plugins:
+  - ../../../../../bundles/@yarnpkg/plugin-licenses.js
