+++ src/__tests__/integration/fixtures/test-package-node-modules/package.json
