+++ src/__tests__/integration/fixtures/test-package-node-modules/packages/package1/package.json
