+++ src/__tests__/integration/fixtures/test-package-pnp/.yarnrc.yml
@@ -1,0 +1,3 @@
+
+plugins:
+  - ../../../../../bundles/@yarnpkg/plugin-licenses.js
