+++ src/__tests__/integration/fixtures/test-package-pnp/package.json
