+++ src/__tests__/integration/fixtures/test-package-pnp/packages/package1/package.json
