+++ src/index.ts
@@ -1,0 +1,8 @@
+import { Plugin } from "@yarnpkg/core";
+import { LicensesListCommand } from "./commands/list";
+
+const plugin: Plugin = {
+  commands: [LicensesListCommand],
+};
+
+export default plugin;
